Attribute Macro msfs_derive::gauge

#[gauge]

Declare a gauge callback. It will be automatically exported with the name NAME_gauge_callback, where NAME is the name of the decorated function.

// Declare and export `FOO_gauge_callback`
#[msfs::gauge]
fn FOO(ctx: &msfs::FsContext, service_id: msfs::PanelServiceID) -> msfs::GaugeCallbackResult {
  // ...
}

The macro can also be given a parameter, name, to rename the exported function.

// Declare and export `FOO_gauge_callback`
#[msfs::gauge(name=FOO)]
fn xyz(...) {}